Created Meetup Event now shows date/time
