+++ react-front-end/package.json
@@ -7,6 +7,7 @@
     "@material-ui/pickers": "^3.3.10",
     "axios": "^0.21.1",
     "formik": "^2.2.6",
+    "moment": "2.29.1",
     "react": "^16.9.0",
     "react-dom": "^16.8.6",
     "react-icons": "^4.2.0",
